Use official OpenJDK image for building
The quarkus one is no longer there, and was only every supposed to be
temporary.

Note that we can't really use the HACBS build definitions bundle for
this directly, as they contain a lot more tasks that we don't want as
part of Java based local dev/e2e testing.