Fix PipelineRun pruning
This fixes a potential race in the pruning of PipelineRun's. Presently these are pruned
before the relevant status update is made on the owner object. Because of the owner reference
this deletion will also update the corresponding owner, making it likely that this update
might fail. Normally failure would result in a requeue but because the oject is question
has been deleted this will no longer work properly.

This fix changes the deletion to happen after the status update, and will also prune any
PipelineRun objects that exist after the object is in its expected state. This means that
even if the prune fails a subsequent reque should allow the prune to succeed.