Handle deployment in the cache
This makes three major changes:
- The cache now handles deployment rather than the sidecar
- Contamination is now handled in a 'smart' manner, where builds are
  only considered contaminated if the contaminated artifacts are
actually requested.
- A new RebuiltArtifact type has been added that represents anything
  that has been rebuilt. This solves an issue where things that were
deployed to the container registry but did not have a corresponding
ArtifactBuild could not be served. This fixes an issue with the Netty
build.

This contamination change means that it is no longer nessesary to fix
contamination issues via ignoring artifacts, and it will prevent
spurious builds from being kicked off that are not actually used by the
main build.