Move to pipeline results driven flows
This works around the results size limit in tekton by directly appending results to the task run using the kube API.

This means we no longer need to create objects directly from the compute workspace inside KCP, but instead the results are observed by the operator and the operator creates the objects directly.

The only objects still needed in the compute workspace is the RebuiltArtifact objects used by the cache for performance reasons, we still need a solution for this.