Make sure ABR state stays in sync with DB state
The DB recieved code was only updating a single owner, however an ABR
can have many owners. For the 'happy path' this was covered by a
reconciliation with state Building which would move the state to
complete.

If the AB had failed due to contamination however only a single AB would
be updated to complete, which is why a lot of the Netty ABs are reported
as failing even though the build has completed.