STONEBLD-847 Initial support for TLS
This will create a TLS service for the Cache, and attempt to configure
all the relevant JVM based tasks to use this route.

This is faily hacky as there is no nice way to integrate the JVM
directly with Kubernetes certificates. At the moment this modifies the
global truststore at the start of the task, which is not ideal, however
there does not appear to be any other way to handle this.

It may be possible to integrate a clustom TrustStore provider into the
JDK that could be configured via environment variables to read
certificates, however this would need some investigation.